python/requirements.txt
# Azure dependencies
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# OpenAI
openai>=1.12.0

# Local vector search
sqlite-vec>=0.1.0

# Data validation
pydantic>=2.5.0

# HTTP requests
requests>=2.31.0

# Model Context Protocol
mcp>=1.0.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0